fix: resolve 4 type conversion warnings in synth_manager and audio_queue
- synth_manager.c: added explicit cast for narrowing conversion (1)
  Line 49: (int)config->sample_rate to convert unsigned enum to int

- audio_queue.c: added explicit casts for multilevel pointer conversions (3)
  Line 45: (const void *)&ptr in xQueueSend
  Line 79: (void *)&ptr in xQueueReceive
  Line 90: (const void *)&ptr in xQueueSend
  Converts uint8_t** to void*/const void* safely

Total: 4 warnings fixed
Build: SUCCESS
Progress: 235/761 warnings (30.9%)

+++ esp_bt_audio_source/components/audio_processor/synth_manager.c
@@ -46,7 +46,7 @@ size_t synth_manager_generate_audio(uint8_t* buffer,
 		return 0;
 	}
 
-	const int sample_rate = (config->sample_rate > 0) ? config->sample_rate : 44100;
+	const int sample_rate = (config->sample_rate > 0) ? (int)config->sample_rate : 44100;
 	const int channels = (config->channels == AUDIO_CHANNEL_MONO) ? 1 : 2;
 	int bytes_per_sample = synth_bytes_per_sample(config->bit_depth);
 	if (bytes_per_sample <= 0) {
